+++ examples/demo_ins_dataset_with_query.py
@@ -69,7 +69,7 @@ def _():
     sys.path.insert(0, str(Path(__file__).parent.parent))
 
     # Use config module for configuration
-    from data_catalog_service.config import get_tiled_url, get_api_key
+    from tiled_catalog_broker.config import get_tiled_url, get_api_key
 
     TILED_URL = get_tiled_url()
     API_KEY = get_api_key()
