+++ tests/test_config.py
@@ -29,7 +29,7 @@ sys.path.insert(0, str(Path(__file__).parent.parent))
 @pytest.fixture(autouse=True)
 def reset_config_cache():
     """Reset config module cache before each test."""
-    import data_catalog_service.config as config
+    import tiled_catalog_broker.config as config
     config._config = None
     yield
     config._config = None
@@ -39,26 +39,26 @@ class TestLoadConfig:
     """Tests for load_config()."""
 
     def test_loads_broker_section(self):
-        from data_catalog_service.config import load_config
+        from tiled_catalog_broker.config import load_config
         cfg = load_config()
         assert isinstance(cfg, dict)
         assert "service_dir" in cfg
 
     def test_no_manifests_section(self):
         """Manifests section removed; code uses fallback pattern."""
-        from data_catalog_service.config import load_config
+        from tiled_catalog_broker.config import load_config
         cfg = load_config()
         assert "manifests" not in cfg
 
     def test_no_dataset_paths(self):
         """The generic config should NOT have hardcoded dataset_paths."""
-        from data_catalog_service.config import load_config
+        from tiled_catalog_broker.config import load_config
         cfg = load_config()
         assert "dataset_paths" not in cfg
 
     def test_no_default_shapes(self):
         """The generic config should NOT have hardcoded default_shapes."""
-        from data_catalog_service.config import load_config
+        from tiled_catalog_broker.config import load_config
         cfg = load_config()
         assert "default_shapes" not in cfg
 
@@ -67,18 +67,18 @@ class TestGetMaxEntities:
     """Tests for get_max_entities()."""
 
     def test_returns_integer(self):
-        from data_catalog_service.config import get_max_entities
+        from tiled_catalog_broker.config import get_max_entities
         result = get_max_entities()
         assert isinstance(result, int)
 
     def test_default_is_positive(self):
-        from data_catalog_service.config import get_max_entities
+        from tiled_catalog_broker.config import get_max_entities
         result = get_max_entities()
         assert result > 0
 
     def test_respects_env_variable(self):
         """Test that MAX_ENTITIES environment variable is respected."""
-        import data_catalog_service.config as config
+        import tiled_catalog_broker.config as config
 
         os.environ["MAX_ENTITIES"] = "42"
         config._config = None
@@ -96,12 +96,12 @@ class TestGetTiledUrl:
     """Tests for get_tiled_url()."""
 
     def test_returns_string(self):
-        from data_catalog_service.config import get_tiled_url
+        from tiled_catalog_broker.config import get_tiled_url
         url = get_tiled_url()
         assert isinstance(url, str)
 
     def test_default_is_localhost(self):
-        from data_catalog_service.config import get_tiled_url
+        from tiled_catalog_broker.config import get_tiled_url
         old_val = os.environ.pop("TILED_URL", None)
         url = get_tiled_url()
         assert "localhost" in url
@@ -109,7 +109,7 @@ class TestGetTiledUrl:
             os.environ["TILED_URL"] = old_val
 
     def test_respects_env_variable(self):
-        from data_catalog_service.config import get_tiled_url
+        from tiled_catalog_broker.config import get_tiled_url
         os.environ["TILED_URL"] = "http://test:9999"
         url = get_tiled_url()
         assert url == "http://test:9999"
@@ -120,12 +120,12 @@ class TestGetApiKey:
     """Tests for get_api_key()."""
 
     def test_returns_string(self):
-        from data_catalog_service.config import get_api_key
+        from tiled_catalog_broker.config import get_api_key
         key = get_api_key()
         assert isinstance(key, str)
 
     def test_default_is_secret(self):
-        from data_catalog_service.config import get_api_key
+        from tiled_catalog_broker.config import get_api_key
         old_val = os.environ.pop("TILED_API_KEY", None)
         key = get_api_key()
         assert key == "secret"
